Move bridge.py from _multi_compartment into _compute
CV/point scatter-gather over CellRuntimeState belongs on the runtime
side. Relocating bridge.py (and its test) removes the back-edge where
_compute/runtime.py imported from _multi_compartment while bridge.py
imported CellRuntimeState back under TYPE_CHECKING, making the
layering _multi_compartment -> _compute -> _discretization strict.

Pure relocation: no function body, signature, or behavior changed.
Updated importers in runtime.py, cell.py, currents.py, probes.py, and
their tests to reference the new location, and rewrote the bridge
module docstrings to describe the module in its new home.
